requirements.txt
# --- Dash y componentes visuales ---
dash==2.17.0
dash-cytoscape==1.0.2
dash-bootstrap-components==1.6.0
plotly==5.22.0

# --- LangChain y NLP ---
langchain==0.2.2
langchain-community==0.2.2
langchain-core==0.2.2

# --- Pinecone y embeddings ---
pinecone-client==3.2.2
openai==1.30.1

# --- Anthropic/Claude ---
requests>=2.31.0

# --- OCR: Tesseract + Pillow + pdf2image ---
pytesseract==0.3.10
Pillow==10.3.0
pdf2image==1.17.0

# --- Utilidades ---
python-dotenv==1.0.1

# --- (Opcional) Flask para g global state (si lo usas) ---
Flask==3.0.3

# --- Otros ---
PyYAML==6.0.1

# --- Pruebas ---
pytest==8.2.1